djacket_django/requirements.txt
djoser # for authentication and tokens
django-cors-headers # Authentication in API and backend
pillow # library for resizing images
stripe
django
django-rest-framework
forex_python
razorpay

